Update Testing and Rename Files
