+++ foundry/foundry.py
@@ -28,7 +28,7 @@ from foundry.models import (
     FoundryDataset
 )
 
-from foundry.eda import (
+from foundry.external_data_architectures import (
     FoundryDataset_Torch
 )
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
@@ -297,7 +297,7 @@ class Foundry(FoundryMetadata):
         """
         return self.dlhub_client.run(name, inputs=inputs, **kwargs)
 
-    def load_data(self, source_id=None, as_hdf5=False):
+    def load_data(self, source_id=None, globus=True, as_hdf5=False):
         """Load in the data associated with the prescribed dataset
 
         Tabular Data Type: Data are arranged in a standard data frame
@@ -325,10 +325,10 @@ class Foundry(FoundryMetadata):
             if self.dataset.splits:
                 for split in self.dataset.splits:
                     data[split.label] = self._load_data(file=split.path,
-                                                        source_id=source_id, as_hdf5=as_hdf5)
+                                                        source_id=source_id, globus=globus, as_hdf5=as_hdf5)
                 return data
             else:
-                return {"data": self._load_data(source_id=source_id, as_hdf5=as_hdf5)}
+                return {"data": self._load_data(source_id=source_id, globus=globus, as_hdf5=as_hdf5)}
         except Exception as e:
             raise Exception(
                 "Metadata not loaded into Foundry object, make sure to call load()") from e
@@ -731,7 +731,7 @@ class Foundry(FoundryMetadata):
             return key_list
 
 
-    def _load_data(self, file=None, source_id=None, as_hdf5=False):
+    def _load_data(self, file=None, source_id=None, globus=True, as_hdf5=False):
 
         # Build the path to access the cached data
         if source_id:
