+++ requirements.txt
@@ -6,7 +6,7 @@ six>=1.11.0
 h5py>=2.10.0
 numpy>=1.15.4
 pandas>=0.23.4
-sklearn>=0.0
+scikit-learn>=1.0
 pydantic>=1.6.1
 joblib>=1.0.0
 mdf_forge>=0.8.0
